+++ locales/ja/auth.json
@@ -14,6 +14,8 @@
     "btn": "変更"
   },
   "password_reset": {
+    "title": "パスワード再設定",
+    "sub_title": "Hapoがパスワード再設定のメールを送信します。",
     "old_password": {
       "label": "現在のパスワード"
     },
@@ -28,6 +30,7 @@
   "validation": {
     "email_required": "メールアドレスは必須です",
     "email_format": "正しいメールアドレスを入力してください",
-    "password_required": "パスワードは必須です"
+    "password_required": "パスワードは必須です",
+    "password_match": "新パスワードと再入力パスワードが一致しません。"
   }
 }
